+++ api/src/main/kotlin/com/few/api/domain/admin/document/usecase/dto/AddArticleUseCaseIn.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.admin.document.dto
+package com.few.api.domain.admin.document.usecase.dto
 
 import java.net.URL
