+++ api/src/main/kotlin/com/few/api/domain/admin/document/usecase/dto/AddWorkbookUseCaseOut.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.admin.document.dto
+package com.few.api.domain.admin.document.usecase.dto
 
 data class AddWorkbookUseCaseOut(
     val workbookId: Long
