+++ api/src/main/kotlin/com/few/api/domain/admin/document/usecase/dto/ConvertContentUseCaseOut.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.admin.document.dto
+package com.few.api.domain.admin.document.usecase.dto
 
 import java.net.URL
