+++ api/src/main/kotlin/com/few/api/domain/article/usecase/dto/ReadArticleUseCaseIn.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.article.dto
+package com.few.api.domain.article.usecase.dto
 
 data class ReadArticleUseCaseIn(
     val articleId: Long
