+++ api/src/main/kotlin/com/few/api/domain/problem/usecase/dto/BrowseProblemsUseCaseIn.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.problem.usecase.`in`
+package com.few.api.domain.problem.usecase.dto
 
 data class BrowseProblemsUseCaseIn(
     val articleId: Long
