+++ api/src/main/kotlin/com/few/api/domain/problem/usecase/dto/BrowseProblemsUseCaseOut.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.problem.usecase.out
+package com.few.api.domain.problem.usecase.dto
 
 data class BrowseProblemsUseCaseOut(
     val problemIds: List<Long>
