+++ api/src/main/kotlin/com/few/api/domain/problem/usecase/dto/ReadProblemUseCaseIn.kt
@@ -1,4 +1,4 @@
-package com.few.api.domain.problem.usecase.`in`
+package com.few.api.domain.problem.usecase.dto
 
 data class ReadProblemUseCaseIn(
     val problemId: Long
