+++ api/src/main/kotlin/com/few/api/domain/workbook/service/WorkbookArticleService.kt
@@ -1,6 +1,6 @@
 package com.few.api.domain.workbook.service
 
-import com.few.api.domain.workbook.dto.ArticleDetail
+import com.few.api.domain.workbook.usecase.dto.ArticleDetail
 import com.few.api.domain.workbook.service.dto.BrowseWorkbookArticlesQuery
 import com.few.api.repo.dao.article.ArticleDao
 import com.few.api.repo.dao.article.query.SelectWorkbookMappedArticleRecordsQuery
