+++ api/src/main/kotlin/com/few/api/web/controller/admin/ApiLogController.kt
@@ -1,7 +1,5 @@
 package com.few.api.web.controller.admin
 
-import com.few.api.domain.admin.document.dto.*
-import com.few.api.domain.admin.document.usecase.*
 import com.few.api.domain.log.AddApiLogUseCase
 import com.few.api.domain.log.dto.AddApiLogUseCaseIn
 import com.few.api.web.controller.admin.request.*
