+++ api/src/test/kotlin/com/few/api/web/controller/problem/ProblemControllerTest.kt
@@ -12,13 +12,13 @@ import com.few.api.web.controller.helper.*
 import com.few.api.web.controller.problem.request.CheckProblemRequest
 import com.few.api.domain.problem.usecase.CheckProblemUseCase
 import com.few.api.domain.problem.usecase.ReadProblemUseCase
-import com.few.api.domain.problem.usecase.`in`.BrowseProblemsUseCaseIn
-import com.few.api.domain.problem.usecase.`in`.CheckProblemUseCaseIn
-import com.few.api.domain.problem.usecase.`in`.ReadProblemUseCaseIn
-import com.few.api.domain.problem.usecase.out.BrowseProblemsUseCaseOut
-import com.few.api.domain.problem.usecase.out.CheckProblemUseCaseOut
-import com.few.api.domain.problem.usecase.out.ReadProblemContentsUseCaseOutDetail
-import com.few.api.domain.problem.usecase.out.ReadProblemUseCaseOut
+import com.few.api.domain.problem.usecase.dto.BrowseProblemsUseCaseIn
+import com.few.api.domain.problem.usecase.dto.CheckProblemUseCaseIn
+import com.few.api.domain.problem.usecase.dto.ReadProblemUseCaseIn
+import com.few.api.domain.problem.usecase.dto.BrowseProblemsUseCaseOut
+import com.few.api.domain.problem.usecase.dto.CheckProblemUseCaseOut
+import com.few.api.domain.problem.usecase.dto.ReadProblemContentsUseCaseOutDetail
+import com.few.api.domain.problem.usecase.dto.ReadProblemUseCaseOut
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
